J is now a class! Define it in a new file!
Overall I realized that having J in a file, in which other stuff
happens, results in errors (J minimizing the wrong function :/).
Having it in a separate class, in separate files, might prevent this.
I used ABCs in order to have a reasonable template and to think of
some properties and functions I want it to have.

I implemented the original J from Massimo, as well as a try of mine.

test_sample.py has been adjusted